Change hologram visibility update interval type from long to int

+++ api/src/main/java/de/oliver/fancyholograms/api/HologramConfiguration.java
@@ -65,5 +65,5 @@ public interface HologramConfiguration {
      *
      * @return The hologram visibility update interval in milliseconds.
      */
-    long getUpdateVisibilityInterval();
+    int getUpdateVisibilityInterval();
 }
